requirements.txt
anthropic>=0.25.0
openpyxl>=3.1.0
python-docx>=1.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0
